Refactored all code, fixed date parsing for covid_19_data.csv
